Address review comments: file rename and error string type change
